/* key와 신고하기 추가 */ admin-newsfeed-story admin-project-feedback admin-project-interview-detail admin-project-user-report
/* nickname 옆에 key만 추가 */
admin-project-home
admin-project-interview

20180116
