+++ src/pages/admin/admin-newsfeed-story/admin-newsfeed-story.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { IonicPage, NavController, NavParams, Slides, App } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, Slides, App, ActionSheetController, AlertController } from 'ionic-angular';
 import { Keyboard } from '@ionic-native/keyboard';
 
 import { CommonServiceProvider } from '../../../providers/common-service/common-service';
@@ -59,7 +59,9 @@ export class AdminNewsfeedStoryPage {
     public appCtrl: App,
     public commonService: CommonServiceProvider,
     public adminService: AdminServiceProvider,
-    public keyboard: Keyboard) {
+    public keyboard: Keyboard,
+    public actionSheetCtrl: ActionSheetController,
+    public alertCtrl: AlertController,) {
   }
 
   ionViewDidLoad() {
@@ -109,6 +111,50 @@ export class AdminNewsfeedStoryPage {
         this.commonService.showBasicAlert('오류가 발생했습니다.');
       }
     );
+  }
+
+  reportContent() {
+    let actionSheet = this.actionSheetCtrl.create({
+      buttons: [
+        {
+          text: '신고하기',
+          role: 'destructive',
+          handler: () => {
+            this.report();
+          }
+        },{
+          text: '취소하기',
+          role: 'cancel',
+          handler: () => {
+            console.log('Cancel clicked');
+          }
+        }
+      ]
+    });
+    actionSheet.present();
+  }
+
+  report() {
+    let alert = this.alertCtrl.create({
+      title: '신고',
+      subTitle: '해당 댓글을 위법/위해<br />댓글로 신고하시겠습니까?',
+      buttons: [
+        {
+          text: '취소',
+          role: 'cancel',
+          handler: data => {
+            console.log('취소');
+          }
+        },
+        {
+          text: '확인',
+          handler: data => {
+            console.log('확인');
+          }
+        }
+      ]
+    });
+    alert.present();
   }
 
   back() {
