+++ src/pages/admin/admin-project-feedback/admin-project-feedback.html
@@ -20,7 +20,10 @@
       </span>
     </ion-title>
     <ion-buttons end>
-      <button ion-button></button>
+      <!-- 피드백 추가  -->
+      <button ion-button (click)="reportContent()">
+        <ion-icon name="md-more"></ion-icon>
+      </button>
     </ion-buttons>
   </ion-navbar>
 </ion-header>
@@ -87,8 +90,18 @@
             <img class="avatar-span-image empathy-image" [src]="opinion.avatar_image" />
           </span>
           <h3>
-            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">{{opinion.nickname}}</span>
-            <span class="empathy-tag opinion-tag">공감</span>
+            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">
+              {{opinion.nickname}}
+              &ensp;<ins class="key-tag">20180116</ins>
+            </span>
+            <span class="opinion-tag-wrapper">
+              <span class="empathy-tag opinion-tag">공감</span>
+              <span class="btn-opinion-more">
+                <button ion-button (click)="reportContent()">
+                  <ion-icon name="md-more"></ion-icon>
+                </button>
+              </span>
+            </span>
           </h3>
           <p class="registration-date" [innerHTML]="opinion.opinion_registration_date | amLocale:'ko' | amTimeAgo"></p>
           <div class="opinion-content">
@@ -103,8 +116,18 @@
             <img class="avatar-span-image non-empathy-image" [src]="opinion.avatar_image" />
           </span>
           <h3>
-            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">{{opinion.nickname}}</span>
-            <span class="non-empathy-tag opinion-tag">비공감</span>
+            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">
+              {{opinion.nickname}}
+              &ensp;<ins class="key-tag">20180116</ins>
+            </span>
+            <span class="opinion-tag-wrapper">
+              <span class="non-empathy-tag opinion-tag">비공감</span>
+              <span class="btn-opinion-more">
+                <button ion-button (click)="reportContent()">
+                  <ion-icon name="md-more"></ion-icon>
+                </button>
+              </span>
+            </span>
           </h3>
           <p class="registration-date" [innerHTML]="opinion.opinion_registration_date | amLocale:'ko' | amTimeAgo"></p>
           <div class="opinion-content">
@@ -128,8 +151,18 @@
             <img class="avatar-span-image empathy-image" [src]="opinion.avatar_image" />
           </span>
           <h3>
-            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">{{opinion.nickname}}</span>
-            <span class="empathy-tag opinion-tag">공감</span>
+            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">
+              {{opinion.nickname}}
+              &ensp;<ins class="key-tag">20180116</ins>
+            </span>
+            <span class="opinion-tag-wrapper">
+              <span class="empathy-tag opinion-tag">공감</span>
+              <span class="btn-opinion-more">
+                <button ion-button (click)="reportContent()">
+                  <ion-icon name="md-more"></ion-icon>
+                </button>
+              </span>
+            </span>
           </h3>
           <p class="registration-date" [innerHTML]="opinion.opinion_registration_date | amLocale:'ko' | amTimeAgo"></p>
           <div class="opinion-content">
@@ -153,8 +186,18 @@
             <img class="avatar-span-image non-empathy-image" [src]="opinion.avatar_image" />
           </span>
           <h3>
-            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">{{opinion.nickname}}</span>
-            <span class="non-empathy-tag opinion-tag">비공감</span>
+            <span class="user-nickname-button" tappable (click)="openAdminProjectUserProfilePage(opinion.project_participant_id)">
+              {{opinion.nickname}}
+              &ensp;<ins class="key-tag">20180116</ins>
+            </span>
+            <span class="opinion-tag-wrapper">
+              <span class="non-empathy-tag opinion-tag">비공감</span>
+              <span class="btn-opinion-more">
+                <button ion-button (click)="reportContent()">
+                  <ion-icon name="md-more"></ion-icon>
+                </button>
+              </span>
+            </span>
           </h3>
           <p class="registration-date" [innerHTML]="opinion.opinion_registration_date | amLocale:'ko' | amTimeAgo"></p>
           <div class="opinion-content">
