+++ src/pages/admin/admin-project-feedback/admin-project-feedback.scss
@@ -193,14 +193,14 @@ page-admin-project-feedback {
         border: 1.25px solid #a8a8a8;
     }
 
-    .opinions .opinion-tag {
-        position: absolute;
-        right: 14px;
-        padding: 2px 13px;
-        font-size: 1.1rem;
-        font-weight: 500;
-        border-radius: 50px;
-    }
+    // .opinions .opinion-tag {
+    //     position: absolute;
+    //     right: 14px;
+    //     padding: 2px 13px;
+    //     font-size: 1.1rem;
+    //     font-weight: 500;
+    //     border-radius: 50px;
+    // }
 
     .opinions .empathy-tag {
         color: #fff;
