+++ src/pages/admin/admin-project-feedback/admin-project-feedback.ts
@@ -1,6 +1,6 @@
 import { Component, ViewChild, ElementRef } from '@angular/core';
 import { SlicePipe } from '@angular/common';
-import { IonicPage, NavController, NavParams, Slides, ModalController, PopoverController } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, Slides, ModalController, PopoverController, ActionSheetController, AlertController } from 'ionic-angular';
 
 import { PhotoViewer } from '@ionic-native/photo-viewer';
 
@@ -53,7 +53,9 @@ export class AdminProjectFeedbackPage {
     public popoverCtrl: PopoverController,
     private photoViewer: PhotoViewer,
     public commonService: CommonServiceProvider,
-    public adminService: AdminServiceProvider) {
+    public adminService: AdminServiceProvider,
+    public actionSheetCtrl: ActionSheetController,
+    public alertCtrl: AlertController,) {
     this.mobWidth = (window.innerWidth);
     this.slideHeight = this.mobWidth * 4 / 5;
   }
@@ -117,6 +119,50 @@ export class AdminProjectFeedbackPage {
     this.commonService.isLoadingActive = true;
     this.ionViewWillEnter();
     refresher.complete();
+  }
+
+  reportContent() {
+    let actionSheet = this.actionSheetCtrl.create({
+      buttons: [
+        {
+          text: '신고하기',
+          role: 'destructive',
+          handler: () => {
+            this.report();
+          }
+        },{
+          text: '취소하기',
+          role: 'cancel',
+          handler: () => {
+            console.log('Cancel clicked');
+          }
+        }
+      ]
+    });
+    actionSheet.present();
+  }
+
+  report() {
+    let alert = this.alertCtrl.create({
+      title: '신고',
+      subTitle: '해당 내용을 위법/위해<br />댓글로 신고하시겠습니까?',
+      buttons: [
+        {
+          text: '취소',
+          role: 'cancel',
+          handler: data => {
+            console.log('취소');
+          }
+        },
+        {
+          text: '확인',
+          handler: data => {
+            console.log('확인');
+          }
+        }
+      ]
+    });
+    alert.present();
   }
 
   back() {
