+++ src/pages/company/company-project-report/company-project-report.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { IonicPage, NavController, NavParams, Slides, ModalController, ActionSheetController, AlertController } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, Slides, ModalController } from 'ionic-angular';
 import { PhotoViewer } from '@ionic-native/photo-viewer';
 
 import { CommonServiceProvider } from '../../../providers/common-service/common-service';
@@ -331,9 +331,7 @@ export class CompanyProjectReportPage {
     public modalCtrl: ModalController,
     public commonService: CommonServiceProvider,
     public companyService: CompanyServiceProvider,
-    public photoViewer: PhotoViewer,
-    public actionSheetCtrl: ActionSheetController,
-    public alertCtrl: AlertController,) {
+    public photoViewer: PhotoViewer) {
   }
 
   ionViewDidLoad() {
@@ -468,51 +466,6 @@ export class CompanyProjectReportPage {
     );
     this.userReportLength = this.userReports.length;
   }
-
-  reportContent() {
-    let actionSheet = this.actionSheetCtrl.create({
-      buttons: [
-        {
-          text: '신고하기',
-          role: 'destructive',
-          handler: () => {
-            this.report();
-          }
-        },{
-          text: '취소하기',
-          role: 'cancel',
-          handler: () => {
-            console.log('Cancel clicked');
-          }
-        }
-      ]
-    });
-    actionSheet.present();
-  }
-
-  report() {
-    let alert = this.alertCtrl.create({
-      title: '신고',
-      subTitle: '해당 내용을 위법/위해<br />댓글로 신고하시겠습니까?',
-      buttons: [
-        {
-          text: '취소',
-          role: 'cancel',
-          handler: data => {
-            console.log('취소');
-          }
-        },
-        {
-          text: '확인',
-          handler: data => {
-            console.log('확인');
-          }
-        }
-      ]
-    });
-    alert.present();
-  }
-
 
   slideChanged() {
     if(this.slides.isBeginning()) {
